Merged PR 8350970: Fix Logical Volume creation
Fix Logical Volume creation:
Description:

In the volume group vg_sap, we need to create several logical volumes which include for lv_sapmnt, lv_usrsapinstall, lv_usrsap. The goal is to allocate a predetermined amount of space to lv_sapmnt, lv_usersap and then allocate the rest of the space to lv_usersapinstall. But the order in which this allocation was being done was in reverse which resulted in lv_usersapinstall consuming all the space in vg_sap therefore, blocking the creation of other logical volumes. The ordering has been fixed in this PR.

This is the test proof for the same:

Query to view the results for createinfra test:

https://waasservices.eastus.kusto.windows.net/WaaSKustoDB?query=H4sIAAAAAAAEAG2RTUsDQQyG7%2fsrwpwqbGFZQUGpILXCHvzAFXoQD2kmtaO7M0sm21Lwxzur0A8qzCFD3id5k8wR6wdWcTRbs9fsGzYrFoZnYXKRX13LtWLbwQ1YVNb0H5VFeT4uLsblJRTF1e8724F1v4gkrlMXfGXBeRhVXlk8NoepuCdeOIZeiJOagld0PoJBuwpkdpq71NfHgXsz8yBfyyZsHrFl8w6TCZipcDJX%2baVgVOlJe%2bH%2f4aeOBQcDaSod8MGgqXsiZsvW5GDu0TUpGgx2Ej6Z9GQb%2bVHNmUiQabD85yiHmmXtiG9J3drptrL5wZBD%2fJHAHPb51CqIZYHF9kh5cgWMdJ1lP74BzVW1AQAA&web=0

WaaSMetricEvent
| where PreciseTimeStamp > datetime(2023-06-27 00:00:00)
| where SubscriptionId in (InternalSubscriptions)
| where ResourceId contains "adhoc"
| where Dimensions["WorkflowName"] == "CreateInfrastructure"
| where Dimensions["OperationState"] in ("Succeeded", "Failed")
| project PreciseTimeStamp, Dimensions["ErrorCodeName"], ServiceActivityId, ResourceId, Region, ActivityId
| order by ResourceId, PreciseTimeStamp asc;

Query to view the results for install test:

https://waasservices.eastus.kusto.windows.net/WaaSKustoDB?query=H4sIAAAAAAAEAG2RTUsDMRCG7%2fsrhpwqbGFZQUGpIFphD37gCj2Ih2ky2mg2WSazWwr%2beBOFflAhhwnv%2b%2bSdmSwQ23sStno%2bkpfiG9YrYoInJm0jvdiOWsGuhyswKCTpPqmr%2bnRanU3rc6iqi99zsgXbYRk1215s8I0B62HSeCH26PaluCOeKYaBNSW3Dl7Q%2bggKzSpotfXcplwfM%2feqFoG%2f3l1YP2BH6g1mM1CNj4LOZcUFNP9zjz0x5uw0kGQy96baQWsiQ0aVoO7QulTl3noOn6TlaBHlwZtz5sA3wdBfMyW0xKPVdK3FjlY2jSn35sv1RwJL2OkpKrAhhuXmwHn0ARj1ZVH8AGqZJGKwAQAA&web=0

WaaSMetricEvent
| where PreciseTimeStamp > datetime(2023-06-27 00:00:00)
| where SubscriptionId in (InternalSubscriptions)
| where ResourceId contains "adhoc"
| where Dimensions["WorkflowName"] == "InstallWorkload"
| where Dimensions["OperationState"] in ("Succeeded", "Failed")
| project PreciseTimeStamp, Dimensions["ErrorCodeName"], ServiceActivityId, ResourceId, Region, ActivityId
| order by ResourceId, PreciseTimeStamp asc;

1 SCS failure observed which was transient and succeeded on retry
These are the other errors observed which are not related to ansible:
InstallHostUnreachable
InstallServiceError
AzureVMIsNotInSupportedProvisioningState

Related work items: #24199283