Fixing Logical Volume creation order (#466)
Fix Logical Volume creation:
Description:

In the volume group vg_sap, we need to create several logical volumes which include for lv_sapmnt, lv_usrsapinstall, lv_usrsap. The goal is to allocate a predetermined amount of space to lv_sapmnt, lv_usersap and then allocate the rest of the space to lv_usersapinstall. But the order in which this allocation was being done was in reverse which resulted in lv_usersapinstall consuming all the space in vg_sap therefore, blocking the creation of other logical volumes. The ordering has been fixed in this PR.

Co-authored-by: Jaskirat Singh <jasksingh@microsoft.com>