Several (bug)fixes for RHEL deployments and deployments in general (#604)
* Add fast_stop=no to pacemaker fileystem resources

Pacemaker isn't respecting the stop timeout on filesystem resources due to the default setting fast_stop=yes.
Without setting fencing will occur because if SAP (A)SCS / ERS isn't stopped in time processes will be terminated
which are restarted by sapstartsrv and node will be fenced because fileystem can't be unmounted.

https://www.suse.com/support/kb/doc/?id=000020860
https://access.redhat.com/solutions/4801371

* Distribute systemd services between SCS / ERS nodes and stop services

Both (A)SCS and ERS systemd services should be present on SCS and ERS nodes otherwise pacemaker only handles
SCS on the SCS node and ERS on the ERS node with the systemd integration.

* Add resource clear for move contrainsts on (A)SCS resource group

* Bugfix folders on local disks to be managed after mounting local disk

* sid_private_key isn't required

* Add sdu_secret prefix/suffix to manage custom Key Vault secret naming